// eqrms2/types/academy-whitepaper-detail.ts
export type AcademyWhitepaperDetail = {
  whitepaper_id: number;
  slug: string | null;
  created_at?: string | null;
  whitepaper_name: string | null;
  whitepaper_date: string | null;
  whitepaper_summary: string | null;
  whitepaper_url: string | null;
  whitepaper_body: string | null;
  whitepaper_img: string | null;
  whitepaper_tagline: string | null;
};